requirements.txt.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
pydantic
pandas
matplotlib
openpyxl
